rpm을 double로 받는다 — 2000.25가 2000으로 잘리고 있었다
백엔드 계약은 OBD-II RPM이 0.25 단위라 Double인데 앱 모델만 int로 받아
fromJson에서 toInt()로 잘랐다. 대시보드는 그 잘린 값을 다시 toDouble()로
넓혀 넘기고 있어서 화면에는 손실이 드러나지 않았다.

표시 동작은 그대로다(원래도 double로 넘어갔다). 계약이 갈라진 것만 맞췄다.

+++ lib/core/models/telemetry.dart
@@ -2,7 +2,7 @@ class Telemetry {
   final String vehicleId;
   final DateTime timestamp;
   final double speed;
-  final int rpm;
+  final double rpm;
   final double engineTemp;
   final double throttlePosition;
   final double fuelLevel;
@@ -30,7 +30,7 @@ class Telemetry {
       vehicleId: json['vehicleId'] as String,
       timestamp: DateTime.parse(json['timestamp'] as String),
       speed: (json['speed'] as num).toDouble(),
-      rpm: (json['rpm'] as num).toInt(),
+      rpm: (json['rpm'] as num).toDouble(),
       engineTemp: (json['engineTemp'] as num).toDouble(),
       throttlePosition: (json['throttlePosition'] as num).toDouble(),
       fuelLevel: (json['fuelLevel'] as num).toDouble(),
